+++ src/url/frag.ts
@@ -1,26 +1,38 @@
-// Sets the document.location.hash while suppressing the viewport scrolling
-// Accepts a plaintext fragment - and URI encodes it automatically - unless the `_isEncoded` flag is set.
-// Usage:
-//   setFrag('myid');
-//   setFrag('#myid');
-//   setFrag('Fraîce 18%');
-//   setFrag('Fra%C3%AEce%2018%25', true);
-//   setFrag('');    // unset
-//   setFrag(null);  // unset
+import getPageScrollElm from '../dom/getPageScrollElm';
 
-// encodes a plain-text string to a URL #fragment friendly format (compatible with .get())
+/**
+ * Encodes a plain-text string to a URL #fragment-friendly
+ * format (compatible with .get())
+ */
 const encodeFrag = (_fragment: string) => {
 	return encodeURI(_fragment).replace(/#/g, '%23').replace(/%7C/g, '|');
 };
 
+/**
+ * Sets the document.location.hash while suppressing the viewport scrolling
+ * Accepts a plaintext fragment - and URI encodes it automatically - unless the `_isEncoded` flag is set.
+ *
+ * Usage:
+ * ```js
+ * setFrag('myid');
+ * setFrag('#myid');
+ * setFrag('Fraîce 18%');
+ * setFrag('Fra%C3%AEce%2018%25', true);
+ * setFrag('');    // unset
+ * setFrag(null);  // unset
+ * ```
+ */
 const setFrag = (_fragment?: string, _isEncoded?: boolean): void => {
 	_fragment = (_fragment || '').replace(/^#/, '');
 	// check if there exists an element with .id same as _fragment
+
+	const scrollElm = getPageScrollElm();
+
 	const _elm =
 		_fragment &&
 		document.getElementById(_isEncoded ? decodeURIComponent(_fragment) : _fragment);
-	const _prePos =
-		window.pageYOffset || document.documentElement.scrollTop || document.body.scrollTop;
+
+	const _prePos = scrollElm.scrollTop;
 	const _tmpId = _elm && _elm.id;
 
 	// temporaily defuse the element's id
@@ -33,14 +45,16 @@ const setFrag = (_fragment?: string, _isEncoded?: boolean): void => {
 	// (because Chrome ~v34 seems to scroll to the element which had -
 	// that ID on page load regardless of wheter its id has changed
 	// or if another element has now received that id. weird.)
-	window.scrollTo(0, _prePos);
+	scrollElm.scrollTo(0, _prePos);
 
 	// put the old DOM id back in it's place
 	_elm && (_elm.id = _tmpId || '');
 };
 
-// returns the #fragment portion of `url` (defaulting to using `document.location.href`)
-// returns a plaintext (decodeURIComponent) version of the fragment - unless a `_returnRaw` argument is passed.
+/**
+ * returns the #fragment portion of `url` (defaulting to using `document.location.href`)
+ * returns a plaintext (decodeURIComponent) version of the fragment - unless a `_returnRaw` argument is passed.
+ */
 const getFrag = (url?: string, _returnRaw?: boolean): string => {
 	const _fragment = (url || document.location.href).split('#')[1] || '';
 	return _returnRaw ? _fragment : decodeURIComponent(_fragment);