+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - ... <!-- Add new lines here. Version number will be decided later -->
 - feat: Add `getPageScrollElm` … returns `<html/>` unless `<body/>` is
   scrollable
+- fix: `setFrag`, `focusElm`, `scrollToElm` assume `<html/>` is always the
+  page-scrolling element
 
 ## 4.7.3